\docType{data}
\name{kegg.list}
\alias{kegg.list}
\title{Kegg Human Annotations Dataframe}
\format{First column as Pathway ID, second column as entrez ID in format of "hsa:XXXXX", 3rd column are human gene symbols and 4th column is name of the pathway.}
\source{
http://www.genome.jp/kegg/
}
\usage{
kegg.list
}
\description{
Kegg Human Annotations Dataframe
}
\author{
Mehran Karimzadeh
}

